requirements.txt
streamlit>=1.28.0
pdfplumber>=0.10.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
